better

+++ src/main/java/com/minecolonies/coremod/entity/pathfinding/PathJobMoveToLocation.java
@@ -6,6 +6,7 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.pathfinding.Path;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.math.Vec3i;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -89,7 +90,7 @@ public class PathJobMoveToLocation extends AbstractPathJob
 
         if (n.pos.getY() == destination.getY() - 1)
         {
-            return destination.withinDistance(new Vec3d(n.pos.getX(), destination.getY(), n.pos.getZ()), DESTINATION_SLACK_ADJACENT);
+            return destination.withinDistance(new Vec3i(n.pos.getX(), destination.getY(), n.pos.getZ()), DESTINATION_SLACK_ADJACENT);
         }
         return destination.withinDistance(n.pos, DESTINATION_SLACK_ADJACENT);
     }
